pyproject.toml
[project]
name = "parsemoneyforward"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "riku.yamada", email = "riku.yamada@optim.co.jp" }
]
dependencies = [
    "selenium>=4.24.0",
    "python-dotenv>=1.0.1",
    "fake-useragent>=1.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/parsemoneyforward"]
